requirements.txt
Django==4.1.2
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
psycopg2==2.9.5
Pillow==9.3.0